feat(credits): wire upgrade-request flow into personal settings
